Add docs and reorganize files/directories. Demote HexagonLayer to sample layer
