+++ docs/layer.md
@@ -1,0 +1,11 @@
+# Layer Class
+
+## Common Parameters
+
+* `id` (string, required): layer ID
+* `viewport` (Viewport, optional, default=injected): viewport
+* `pickable` [bool, optional, default=false] whether layer responds to
+   mouse events
+* `visible` [bool, optional, default=true] whether layer is drawn
+
+* `opacity` (number, required) opacity of the layer
