+++ src/layers/core/index.js
@@ -1,6 +1,5 @@
-export {default as HexagonLayer} from './hexagon-layer';
-export {default as ChoroplethLayer} from './choropleth-layer';
-export {default as ScatterplotLayer} from './scatterplot-layer';
-export {default as GridLayer} from './grid-layer';
 export {default as ArcLayer} from './arc-layer';
+export {default as ChoroplethLayer} from './choropleth-layer';
+export {default as GridLayer} from './grid-layer';
 export {default as LineLayer} from './line-layer';
+export {default as ScatterplotLayer} from './scatterplot-layer';
