--- src/layers/experimental/index.js
@@ -1,13 +1,0 @@
-import EnhancedHexagonLayer from './enhanced-hexagon-layer';
-import ExtrudedChoroplethLayer from './extruded-choropleth-layer';
-import PointCloudLayer from './point-cloud-layer';
-
-import ExperimentalScatterplotLayer from './scatterplot-layer';
-
-module.exports = {
-  EnhancedHexagonLayer,
-  ExtrudedChoroplethLayer,
-  PointCloudLayer,
-
-  ExperimentalScatterplotLayer
-};
