+++ src/layers/fp64/index.js
@@ -1,5 +1,1 @@
-import ScatterplotLayer64 from './scatterplot-layer';
-
-module.exports = {
-  ScatterplotLayer64
-};
+export {default as ScatterplotLayer64} from './scatterplot-layer';
