+++ src/layers/samples/arc-layer/arc-layer-fragment.js
