+++ src/layers/samples/enhanced-hexagon-layer/enhanced-hexagon-layer.js
