+++ src/layers/samples/enhanced-hexagon-layer/index.js
