+++ src/layers/samples/extruded-choropleth-layer/extruded-choropleth-layer-fragment.js
