+++ src/layers/samples/extruded-choropleth-layer/extruded-choropleth-layer-vertex.js
