+++ src/layers/samples/extruded-choropleth-layer/extruded-choropleth-layer.js
