+++ src/layers/samples/hexagon-layer/hexagon-layer-fragment.glsl
