+++ src/layers/samples/hexagon-layer/hexagon-layer.js
