+++ src/layers/samples/hexagon-layer/index.js
