+++ src/layers/samples/index.js
@@ -1,0 +1,5 @@
+export {default as HexagonLayer} from './hexagon-layer';
+export {default as EnhancedHexagonLayer} from './enhanced-hexagon-layer';
+export {default as ExtrudedChoroplethLayer} from './extruded-choropleth-layer';
+export {default as PointCloudLayer} from './point-cloud-layer';
+export {default as ExperimentalScatterplotLayer} from './scatterplot-layer';
