+++ src/layers/samples/point-cloud-layer/index.js
