+++ src/layers/samples/point-cloud-layer/point-cloud-layer-vertex.js
