+++ src/layers/samples/scatterplot-layer/scatterplot-layer-fragment.js
