+++ src/layers/samples/scatterplot-layer/scatterplot-layer-vertex.js
