+++ src/layers/samples/scatterplot-layer/scatterplot-layer.js
