--- src/layers/test/index.js
@@ -1,2 +1,0 @@
-export {default as TestLayer} from './test-layer';
-export {default as TriangleLayer} from './triangle-layer';
