+++ src/test/browser.js
@@ -2,4 +2,4 @@ require('tap-browser-color')();
 require('babel-polyfill');
 
 require('./');
-require('./deckgl-overlay-spec');
+require('./react');
