+++ src/test/electron.js
@@ -1,4 +1,4 @@
 require('babel-polyfill');
 
 require('./');
-require('./deckgl-overlay-spec');
+require('./react');
