+++ src/test/imports-spec.js
@@ -1,24 +1,23 @@
 import test from 'tape-catch';
 import 'luma.gl/headless';
 import {
-  DeckGLOverlay, Layer, BaseLayer,
-  HexagonLayer, ChoroplethLayer, ScatterplotLayer,
+  DeckGL, Layer, BaseLayer,
+  ChoroplethLayer, ScatterplotLayer,
   GridLayer, ArcLayer, LineLayer,
   Viewport
 } from '../..';
 
 import {
-  PointCloudLayer, ExtrudedChoroplethLayer, EnhancedHexagonLayer
-} from '../../experimental';
+  PointCloudLayer, ExtrudedChoroplethLayer, HexagonLayer, EnhancedHexagonLayer
+} from '../../samples';
 
 import Viewport2 from '../../viewport';
 
 test('Top-level imports', t0 => {
   t0.test('import "deck.gl"', t => {
-    t.ok(DeckGLOverlay, 'DeckGLOverlay symbol imported');
+    t.ok(DeckGL, 'DeckGLOverlay symbol imported');
     t.ok(Layer, 'Layer symbol imported');
     t.ok(BaseLayer, 'BaseLayer symbol imported');
-    t.ok(HexagonLayer, 'HexagonLayer symbol imported');
     t.ok(ChoroplethLayer, 'ChoroplethLayer symbol imported');
     t.ok(ScatterplotLayer, 'ScatterplotLayer symbol imported');
     t.ok(GridLayer, 'GridLayer symbol imported');
@@ -28,9 +27,10 @@ test('Top-level imports', t0 => {
     t.end();
   });
 
-  t0.test('import "deck.gl/experimental"', t => {
+  t0.test('import "deck.gl/samples"', t => {
     t.ok(PointCloudLayer, 'PointCloudLayer symbol imported');
     t.ok(ExtrudedChoroplethLayer, 'ExtrudedChoroplethLayer symbol imported');
+    t.ok(HexagonLayer, 'HexagonLayer symbol imported');
     t.ok(EnhancedHexagonLayer, 'EnhancedHexagonLayer symbol imported');
     t.end();
   });
