+++ src/test/layers/index.js
@@ -1,0 +1,1 @@
+import './core-layers.spec';
