+++ src/test/lib/base-layer.spec.js
