+++ src/test/react/index.js
@@ -1,0 +1,1 @@
+import './deckgl.spec';
