+++ src/test/viewport/index.js
@@ -1,0 +1,1 @@
+import './viewport.spec';
